% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{isValidEmail}
\alias{isValidEmail}
\title{is Valid Email}
\usage{
isValidEmail(email)
}
\arguments{
\item{email}{email entered by user}
}
\description{
Help validate email entered by user
}
\keyword{internal}
